% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_column.R
\name{pull_column}
\alias{pull_column}
\title{Pulls values of select data frame column}
\usage{
pull_column(data, column_name)
}
\arguments{
\item{data}{Data frame}

\item{column_name}{Column name to pull, in quotes}
}
\value{
A vector
}
\description{
Pulls values of select data frame column
}
\examples{
pull_column(iris, "Sepal.Width")

}
